refactoring all fallback logic to new ApiClientFallback class
