+++ blibli-backend-framework-api-client/src/main/java/com/blibli/oss/backend/apiclient/aop/fallback/FallbackMetadata.java
@@ -1,4 +1,4 @@
-package com.blibli.oss.backend.apiclient.aop;
+package com.blibli.oss.backend.apiclient.aop.fallback;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
