+++ blibli-backend-framework-api-client/src/main/java/com/blibli/oss/backend/apiclient/aop/fallback/FallbackMetadataBuilder.java
@@ -1,4 +1,4 @@
-package com.blibli.oss.backend.apiclient.aop;
+package com.blibli.oss.backend.apiclient.aop.fallback;
 
 import org.springframework.util.ReflectionUtils;
